Fix flaky agentprotocol test
Both testing threads called the Kill() function at the same time which
caused them to deadlock as they both were trying to send the termination
packet to each-other at the same time. In a real situation this would
not cause a deadlock as the communication between the two sides would be
buffered.

Signed-off-by: Nikos Tsipinakis <nikos@tsipinakis.com>